Modify test case, handle space after comma in whiteboard, add =, rename file
