+++ src/utils/extractWhiteboardTags.js
@@ -1,23 +1,30 @@
-const extractWhiteboardTags = whiteboard => {
-  // find occurence like lang=perl,js or lang=perl/js
-  const re = /(\w+)=(\w+)[,|/](\w+)/;
+export default whiteboard => {
+  // Find whiteboard tags of the form:
+  // [lang=js][lang=xul]
+  // [lang=js,xul]
+  // [lang=js/xul]
+  if (!whiteboard) {
+    return [];
+  }
+
+  const re = /(\w+)=(\w+)[,|/]\s?(\w+)/;
   // find all element within square bracket starting with lang with as little
   // string as possible inside
-  const keywords = (whiteboard.match(/\[lang.+?\]/g) || []).reduce(
-    (prev, item) => {
-      // remove the square bracket
+  const keywords = (whiteboard.match(/\[lang=.+?\]/g) || []).reduce(
+    (acc, item) => {
+      // remove the square brackets
       const itemWithoutBracket = item.slice(1, -1);
       const items = itemWithoutBracket.match(re);
 
       // if there is no comma or /
       if (!items) {
-        return [...prev, itemWithoutBracket];
+        return [...acc, itemWithoutBracket];
       }
 
       // items=["lang=js/xul","lang","js","xul"]
       // so start looping from 3rd element
       return [
-        ...prev,
+        ...acc,
         ...items.slice(2).map(element => `${items[1]}=${element}`),
       ];
     },
@@ -26,5 +33,3 @@ const extractWhiteboardTags = whiteboard => {
 
   return keywords;
 };
-
-export default extractWhiteboardTags;
