+++ test/extract_whiteboard_test.js
@@ -1,24 +1,24 @@
-import extractWhiteboardTags from '../src/utils/extractWhiteboard';
+import extractWhiteboardTags from '../src/utils/extractWhiteboardTags';
 
 describe('extract whiteboard tags', () => {
-  it('should get the language', () => {
-    const whiteboard = '[lang=js][lang=xul]';
-    const result = extractWhiteboardTags(whiteboard);
+  it('should extract languages from [lang=js][lang=xul]', () => {
+    const result = extractWhiteboardTags('[lang=js][lang=xul]');
     expect(result).toEqual(['lang=js', 'lang=xul']);
   });
-  it('should ignore other whiteboard tag besides language', () => {
-    const whiteboard = '[good-first-bug]';
-    const result = extractWhiteboardTags(whiteboard);
-    expect(result).toEqual([]);
-  });
-  it('should handle comma tag', () => {
-    const whiteboard = '[lang=js,xul]';
-    const result = extractWhiteboardTags(whiteboard);
+  it('should extract languages from [lang=js][lang=xul][good-first-bug]', () => {
+    const result = extractWhiteboardTags('[lang=js][lang=xul][good-first-bug]');
     expect(result).toEqual(['lang=js', 'lang=xul']);
   });
-  it('should handle OR tag', () => {
-    const whiteboard = '[lang=js/xul]';
-    const result = extractWhiteboardTags(whiteboard);
+  it('should extract languages from [lang=js,xul]', () => {
+    const result = extractWhiteboardTags('[lang=js,xul]');
+    expect(result).toEqual(['lang=js', 'lang=xul']);
+  });
+  it('should extract languages from [lang=js/xul]', () => {
+    const result = extractWhiteboardTags('[lang=js/xul]');
+    expect(result).toEqual(['lang=js', 'lang=xul']);
+  });
+  it('should extract languages from [lang=js, xul]', () => {
+    const result = extractWhiteboardTags('[lang=js, xul]');
     expect(result).toEqual(['lang=js', 'lang=xul']);
   });
 });
